Update code.py
added web-page for debug printing. to be used when running without usb. the better method would be to send stdout and stderr to the web console instead of just whatever args i send to `debugger`. this should do for now though since i believe the error is happening in my exceptions